Restructured project and fixed imports
