+++ .gitignore
@@ -1,1 +1,1 @@
-config.py
+BOT/config.py
